refactor: send keys to the parser
This allows us to decouple key normalisation from the loader to the
parser. The parser can decide how to format the key and maintain
a mapping to the original key. This is useful for parsers that are case
insensitive like environment variables. The parser now must decicde how
to handle those scenarios. Most will not need to, e.g the file based
parsers like json,toml and yaml.